Create 489C - Given Length and Sum of Digits....cpp
Output Status: 

By Rajakunalpandit, contest: Codeforces Round #277.5 (Div. 2), problem: (C) Given Length and Sum of Digits..., Accepted

You have a positive integer m and a non-negative integer s. Your task is to find the smallest and the largest of the numbers that have length m and sum of digits s. The required numbers should be non-negative integers written in the decimal base without leading zeroes.

Input
The single line of the input contains a pair of integers m, s (1 ≤ m ≤ 100, 0 ≤ s ≤ 900) — the length and the sum of the digits of the required numbers.

Output
In the output print the pair of the required non-negative integer numbers — first the minimum possible number, then — the maximum possible number. If no numbers satisfying conditions required exist, print the pair of numbers "-1 -1" (without the quotes).

Examples
inputCopy
2 15
outputCopy
69 96
inputCopy
3 0
outputCopy
-1 -1